file structure, made get_power_2 visible
